refactor(core): co-locate RuntimeError-related scipts
This commit moves some scripts around to co-locate RuntimeError-related scipts under `packages/core/src/errors` folder.
