+++ packages/core/src/errors/error_details_base_url.ts
