+++ packages/core/src/errors/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ERROR_DETAILS_PAGE_BASE_URL} from './error_details_base_url';
+import {createRuntimeErrorClass} from './util';
 
 /**
  * The list of error codes used in runtime code.
@@ -57,40 +57,6 @@ export const enum PackageErrorPrefix {
   FORMS = 1,
 }
 
-export interface RuntimeError<T> {
-  code: T;
-}
-
-export interface RuntimeErrorCtor<T> {
-  new(code: T, message: string): RuntimeError<T>;
-}
-
-export function createRuntimeErrorClass<T = RuntimeErrorCode>(packagePrefix: PackageErrorPrefix):
-    RuntimeErrorCtor<T> {
-  class RuntimeErrorImpl<T = RuntimeErrorCode> extends Error {
-    constructor(public code: T, message: string) {
-      super(formatRuntimeError<T>(code, message, packagePrefix));
-    }
-  }
-  return RuntimeErrorImpl as unknown as RuntimeErrorCtor<T>;
-}
-
+// Main class that should be used to throw runtime errors in the core package
 // tslint:disable-next-line:no-toplevel-property-access
 export const RuntimeError = createRuntimeErrorClass(PackageErrorPrefix.CORE);
-
-/** Called to format a runtime error */
-export function formatRuntimeError<T = RuntimeErrorCode>(
-    code: T, message: string, packagePrefix: PackageErrorPrefix = PackageErrorPrefix.CORE): string {
-  const codeAsNumber = code as unknown as number;
-  // Error code might be a negative number, which is a special marker that instructs the logic to
-  // generate a link to the error details page on angular.io.
-  const fullCode = `NG${packagePrefix}${Math.abs(codeAsNumber)}`;
-
-  // TODO: add a test case for an empty `message` (make sure only code is produced).
-  let errorMessage = `${fullCode}${message ? ': ' + message : ''}`;
-
-  if (ngDevMode && codeAsNumber < 0) {
-    errorMessage = `${errorMessage}. Find more at ${ERROR_DETAILS_PAGE_BASE_URL}/${fullCode}`;
-  }
-  return errorMessage;
-}
