+++ packages/core/src/render3/errors_di.ts
@@ -6,8 +6,9 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {InjectorType} from '../di/interface/defs';
+import {RuntimeError, RuntimeErrorCode} from '../errors';
 import {stringify} from '../util/stringify';
-import {RuntimeError, RuntimeErrorCode} from './error_code';
+
 import {stringifyForError} from './util/stringify_utils';
