+++ packages/core/test/di/reflective_injector_spec.ts
@@ -9,7 +9,7 @@
 import {forwardRef, Inject, Injectable, InjectionToken, Injector, Optional, Provider, ReflectiveInjector, ReflectiveKey, Self} from '@angular/core';
 import {ReflectiveInjector_} from '@angular/core/src/di/reflective_injector';
 import {ResolvedReflectiveProvider_} from '@angular/core/src/di/reflective_provider';
-import {getOriginalError} from '@angular/core/src/errors';
+import {getOriginalError} from '@angular/core/src/util/errors';
 import {expect} from '@angular/platform-browser/testing/src/matchers';
 
 import {stringify} from '../../src/util/stringify';
